First cut at a generic KioskWindowManager
