+++ examples/example-server-lib/kiosk_window_manager_policy.h
@@ -1,0 +1,49 @@
+/*
+ * Copyright © Canonical Ltd.
+ *
+ * This program is free software: you can redistribute it and/or modify it
+ * under the terms of the GNU General Public License version 2 or 3 as
+ * published by the Free Software Foundation.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ */
+
+#ifndef EXAMPLE_KIOSK_WINDOW_MANAGER_POLICY_H
+#define EXAMPLE_KIOSK_WINDOW_MANAGER_POLICY_H
+
+#include <miral/canonical_window_manager.h>
+
+#include <mir_toolkit/events/enums.h>
+
+using namespace mir::geometry;
+
+/// Base class for kiosk-style window managers with common input handling
+/// Subclasses pass the target window state (fullscreen or maximized) to the constructor.
+class KioskWindowManagerPolicyBase : public miral::CanonicalWindowManagerPolicy
+{
+public:
+    KioskWindowManagerPolicyBase(miral::WindowManagerTools const& tools, MirWindowState target_state);
+
+    auto place_new_window(miral::ApplicationInfo const& app_info, miral::WindowSpecification const& request)
+    -> miral::WindowSpecification override;
+
+    bool handle_keyboard_event(MirKeyboardEvent const* event) override;
+    bool handle_touch_event(MirTouchEvent const* event) override;
+    bool handle_pointer_event(MirPointerEvent const* event) override;
+    void handle_modify_window(miral::WindowInfo& window_info, miral::WindowSpecification const& modifications) override;
+
+    void handle_request_move(miral::WindowInfo& window_info, MirInputEvent const* input_event) override;
+    void handle_request_resize(miral::WindowInfo& window_info, MirInputEvent const* input_event,
+                               MirResizeEdge edge) override;
+
+protected:
+    MirWindowState const target_state;
+};
+
+#endif /* EXAMPLE_KIOSK_WINDOW_MANAGER_POLICY_H */
