+++ examples/mir-x11-kiosk/CMakeLists.txt
@@ -3,7 +3,7 @@ mir_add_wrapped_executable(mir-x11-kiosk
     x11_kiosk_window_manager.cpp x11_kiosk_window_manager.h
 )
 
-target_link_libraries(mir-x11-kiosk PRIVATE miral)
+target_link_libraries(mir-x11-kiosk PRIVATE miral example-shell-lib)
 
 add_custom_target(mir-x11-kiosk-launch ALL
     cp ${CMAKE_CURRENT_SOURCE_DIR}/mir-x11-kiosk-launch.sh ${CMAKE_RUNTIME_OUTPUT_DIRECTORY}/mir-x11-kiosk-launch
