+++ examples/mir-x11-kiosk/x11_kiosk_window_manager.cpp
@@ -16,76 +16,15 @@
 
 #include "x11_kiosk_window_manager.h"
 
-#include <miral/application_info.h>
 #include <miral/toolkit_event.h>
 #include <miral/window_info.h>
 #include <miral/window_manager_tools.h>
 
-#include <linux/input.h>
-
-namespace ms = mir::scene;
 using namespace miral;
-using namespace miral::toolkit;
 
-bool X11KioskWindowManagerPolicy::handle_keyboard_event(MirKeyboardEvent const* /*event*/)
+X11KioskWindowManagerPolicy::X11KioskWindowManagerPolicy(WindowManagerTools const& tools) :
+    KioskWindowManagerPolicyBase{tools, mir_window_state_fullscreen}
 {
-    return false;
-}
-
-bool X11KioskWindowManagerPolicy::handle_touch_event(MirTouchEvent const* event)
-{
-    auto const count = mir_touch_event_point_count(event);
-
-    long total_x = 0;
-    long total_y = 0;
-
-    for (auto i = 0U; i != count; ++i)
-    {
-        total_x += mir_touch_event_axis_value(event, i, mir_touch_axis_x);
-        total_y += mir_touch_event_axis_value(event, i, mir_touch_axis_y);
-    }
-
-    Point const cursor{total_x/count, total_y/count};
-
-    tools.select_active_window(tools.window_at(cursor));
-
-    return false;
-}
-
-bool X11KioskWindowManagerPolicy::handle_pointer_event(MirPointerEvent const* event)
-{
-    auto const action = mir_pointer_event_action(event);
-
-    Point const cursor{
-        mir_pointer_event_axis_value(event, mir_pointer_axis_x),
-        mir_pointer_event_axis_value(event, mir_pointer_axis_y)};
-
-    if (action == mir_pointer_action_button_down)
-    {
-        tools.select_active_window(tools.window_at(cursor));
-    }
-
-    return false;
-}
-
-auto X11KioskWindowManagerPolicy::place_new_window(ApplicationInfo const& app_info, WindowSpecification const& request)
--> WindowSpecification
-{
-    WindowSpecification specification = CanonicalWindowManagerPolicy::place_new_window(app_info, request);
-
-    if ((specification.type() == mir_window_type_normal || specification.type() == mir_window_type_freestyle) &&
-        (!specification.parent().has_value() || !specification.parent().value().lock()))
-    {
-        specification.state() = mir_window_state_fullscreen;
-        specification.size() = std::nullopt; // Ignore requested size (if any) when we fullscreen
-        specification.top_left() = std::nullopt; // Ignore requested position (if any) when we fullscreen
-        tools.place_and_size_for_state(specification, WindowInfo{});
-
-        if (!request.state().has_value() || request.state().value() != mir_window_state_restored)
-            specification.state() = request.state();
-    }
-
-    return specification;
 }
 
 void X11KioskWindowManagerPolicy::handle_modify_window(WindowInfo& window_info, WindowSpecification const& modifications)
@@ -124,18 +63,10 @@ void X11KioskWindowManagerPolicy::handle_modify_window(WindowInfo& window_info, 
         }
     }
 
-    CanonicalWindowManagerPolicy::handle_modify_window(window_info, specification);
+    KioskWindowManagerPolicyBase::handle_modify_window(window_info, specification);
 }
 
-void X11KioskWindowManagerPolicy::handle_request_move(WindowInfo& /*window_info*/, MirInputEvent const* /*input_event*/)
-{
-}
-
-void X11KioskWindowManagerPolicy::handle_request_resize(WindowInfo& /*window_info*/, MirInputEvent const* /*input_event*/, MirResizeEdge /*edge*/)
-{
-}
-
-void X11KioskWindowManagerPolicy::handle_window_ready(miral::WindowInfo& window_info)
+void X11KioskWindowManagerPolicy::handle_window_ready(WindowInfo& window_info)
 {
     if ((window_info.type() == mir_window_type_normal || window_info.type() == mir_window_type_freestyle) &&
         !window_info.parent())
@@ -150,5 +81,5 @@ void X11KioskWindowManagerPolicy::handle_window_ready(miral::WindowInfo& window_
             tools.modify_window(window_info, specification);
         }
     }
-    CanonicalWindowManagerPolicy::handle_window_ready(window_info);
+    KioskWindowManagerPolicyBase::handle_window_ready(window_info);
 }
