+++ examples/mir-x11-kiosk/x11_kiosk_window_manager.h
@@ -17,29 +17,15 @@
 #ifndef MIRAL_X11_KIOSK_WINDOW_MANAGER_H
 #define MIRAL_X11_KIOSK_WINDOW_MANAGER_H
 
-#include <miral/canonical_window_manager.h>
+#include "kiosk_window_manager_policy.h"
 
-#include <mir_toolkit/events/enums.h>
-
-using namespace mir::geometry;
-
-class X11KioskWindowManagerPolicy : public miral::CanonicalWindowManagerPolicy
+class X11KioskWindowManagerPolicy : public KioskWindowManagerPolicyBase
 {
 public:
-    using miral::CanonicalWindowManagerPolicy::CanonicalWindowManagerPolicy;
+    X11KioskWindowManagerPolicy(miral::WindowManagerTools const& tools);
 
-    auto place_new_window(miral::ApplicationInfo const& app_info, miral::WindowSpecification const& request)
-    -> miral::WindowSpecification override;
-
-    bool handle_keyboard_event(MirKeyboardEvent const* event) override;
-    bool handle_touch_event(MirTouchEvent const* event) override;
-    bool handle_pointer_event(MirPointerEvent const* event) override;
     void handle_modify_window(miral::WindowInfo& window_info, miral::WindowSpecification const& modifications) override;
-
-    void handle_request_move(miral::WindowInfo& window_info, MirInputEvent const* input_event) override;
     void handle_window_ready(miral::WindowInfo& window_info) override;
-    void handle_request_resize(miral::WindowInfo& window_info, MirInputEvent const* input_event,
-                               MirResizeEdge edge) override;
 };
 
 #endif /* MIRAL_X11_KIOSK_WINDOW_MANAGER_H */
